+++ src/main/java/com/fravokados/mindim/block/IFacingSix.java
@@ -4,6 +4,6 @@ package com.fravokados.mindim.block;
  * @author Nuklearwurst
  */
 public interface IFacingSix {
-	void setFacing(byte b);
-	byte getFacing();
+	void setFacing(short b);
+	short getFacing();
 }
